+++ src/tac_fuse/training/__init__.py
@@ -3,6 +3,7 @@
 from tac_fuse.training.hpo_config import default_hpo_config, should_schedule_hpo
 from tac_fuse.training.model_package import (
     DEFAULT_SIGLIP2_CLASSIFIER_PACKAGE,
+    inspect_packaged_siglip2_package,
     load_siglip2_classifier_package,
     packaged_siglip2_model_dir,
     packaged_siglip2_required_files,
@@ -13,6 +14,7 @@ __all__ = [
     "DEFAULT_SIGLIP2_CLASSIFIER_PACKAGE",
     "SigLIP2INT8Config",
     "default_hpo_config",
+    "inspect_packaged_siglip2_package",
     "load_siglip2_classifier_package",
     "packaged_siglip2_model_dir",
     "packaged_siglip2_required_files",